Add conditional check to before purging pkg directory.

+++ scripts/build-release.sh
@@ -2,7 +2,10 @@
 set -e -x
 # Cross-compile for all platforms with gox
 
-rm -r pkg/*
+# Clean old builds if present.
+if [ -d pkg ]; then
+    rm -r pkg/*
+fi
 
 PROJECT_NAME=${PWD##*/}
 RELEASE_VERSION="$(git describe --abbrev=0)"
